// src/shared/api/generated/models/examinationManageContractDtoSubjectVMockExam.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * i车保WebApi中间件
 * OpenAPI spec version: v1
 */

export interface ExaminationManageContractDtoSubjectVMockExam {
  /** 考试名称 */
  examTypeName?: string;
  /** 试题数量 */
  subjectCount?: number;
  /** 考试题型分布 */
  subjectTypeDistribute?: string;
  /** 考试说明 */
  examInterpretation?: string;
  /** 时间 */
  time?: number;
}