// src/shared/api/generated/models/examinationManageContractDtoSubjectSubjectDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * i车保WebApi中间件
 * OpenAPI spec version: v1
 */
import type { ExaminationManageContractDtoSubjectLabelTemplate } from './examinationManageContractDtoSubjectLabelTemplate';

export interface ExaminationManageContractDtoSubjectSubjectDto {
  id?: string;
  title?: string;
  selection?: string;
  desc?: string;
  answer?: string;
  /** 难易程度 */
  easyLevel?: number;
  img?: string;
  skillDesc?: string;
  videoUrl?: string;
  videoPic?: string;
  score?: number;
  remark?: string;
  examTypeId?: string;
  examTypeName?: string;
  examTypeParentId?: string;
  examTypeParentName?: string;
  type?: string;
  typeName?: string;
  labels?: ExaminationManageContractDtoSubjectLabelTemplate;
  number?: number;
  /** 想法数量 */
  notesCount?: number;
  isCollection?: boolean;
}