// src/shared/api/generated/models/examinationManageContractDtoSubjectVGradeHistory.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * i车保WebApi中间件
 * OpenAPI spec version: v1
 */
import type { ExaminationManageContractDtoSubjectGradeHistoryDto } from './examinationManageContractDtoSubjectGradeHistoryDto';

export interface ExaminationManageContractDtoSubjectVGradeHistory {
  /** 刷题率 */
  answerRate?: string;
  /** 考试名称 */
  examTypeName?: string;
  /** 合格次数 */
  passTimes?: number;
  /** 考试次数 */
  examTimes?: number;
  /** 平均分 */
  averageScore?: number;
  /** 最高分 */
  highestScore?: number;
  /** 最低分 */
  lowestScore?: number;
  gradeHistories?: ExaminationManageContractDtoSubjectGradeHistoryDto[];
}