// src/shared/api/generated/models/examinationManageContractDtoSubjectSubmitExerciseRecordInput.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * i车保WebApi中间件
 * OpenAPI spec version: v1
 */
import type { ExaminationManageContractDtoSubjectSubjectBase } from './examinationManageContractDtoSubjectSubjectBase';

export interface ExaminationManageContractDtoSubjectSubmitExerciseRecordInput {
  examTypeId?: string;
  /** 套题ID */
  subjectGroupId?: string;
  subjectGroupType?: number;
  subjectList?: ExaminationManageContractDtoSubjectSubjectBase[];
  /** 耗费时间(单位秒) */
  time?: number;
}