// src/shared/api/generated/models/examinationManageContractDtoSubjectSubjectBase.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * i车保WebApi中间件
 * OpenAPI spec version: v1
 */

export interface ExaminationManageContractDtoSubjectSubjectBase {
  id?: string;
  /** 是否错误 */
  isMistake?: boolean;
  /** 是否收藏 */
  isCollection?: boolean;
  /** 用户的回答 */
  answer?: string;
}